pyproject.toml
[tool.poetry]
name = "se-interview"
version = "0.1.0"
description = "A simple LangGraph agent with web search capability"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.11,<3.14"
langgraph = "^0.2"
langchain = "^0.3"
langchain-community = "^0.3"
langchain-openai = "^0.2"
ddgs = "^9.0"
python-dotenv = "^1.0"
fastapi = "^0.115"
uvicorn = "^0.32"
google-search-results = "^2.4.2"
openinference-instrumentation-langchain = "^0.1.61"
opentelemetry-exporter-otlp-proto-http = "^1.40.0"
arize-phoenix-otel = "^0.15.0"
arize-phoenix-evals = "^2.12.0"
arize-phoenix-client = "^2.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
